Added new project NativeCore
The idea behind is to put there all the shared functionalities to
BlackBerry Native SDK. It will not have any references to the UI etc and
will be much easier to test and reuse. Potentially it could also be a
place for the new C# versio of the GDB command parser.