Temporarly loading private-debug version of DebugEngine
Only for a time of refactoring all over to remove the C++ code dependency.
Normally the BlackBerry.DebugEngine.dll depends on the
BlackBerry.Package.dll and GDBParser.dll, while it should be reverted. The
Package should reference DebugEngine to have it automatically included
inside the VSIX-package. GDBParser should be rewritten to C# and placed
inside NativeCore library. This will eliminate the artificial x86 code
enforcement.