Improved processing of notifications from GDB
Now parsed content is passed to the DE via a dedicated thread (as in C++
GDBParser) not to block communication with GDB. Fixed also a bug, that
could case a hang (infinite loop) of paring-instruction, when processing
'breakpoint-modified' notification. Finally added more tests.

Unfurtunatelly there are still actions in VisualStudio that could case an
improper behaviour:
 * stepping in source-code
 * notifications can start coming before DE is fully initialized
 * complex breakpoints edition can cause unexpected by BreakpointManager
   responses from GDB