Aligned VS2010 MSBuild with VS2012
The same project can now be opened in one or another version of
VisualStudio and be build without major issues. The same variables and
behaviours are implemented.

NOTE: Looks like VS2010 has a limitation and can not update default
platform values ad hoc. That's why $(OutDir) / $(IntDir) might be
different, when using both VS simultanously. The workaround here is to
change the paths in project settings using $(TargetCPU)-$(Configuration),
that will solve it instantly.