Unified flag files management
Fixed the way -debug-native & CSK-passwords are passed from Visual Studio
to MSBuild. The idea is to use files in project's folder, automatically
deleted, when build/deployment is done.

This commit also fixes non-working debug-native flag, so it is possible now to
put and stop on GDB breakpoint inside the first line of main() method for desired application.

CSK-password is used to pass a CSK certyficate key used during BAR signing
process (in case someone doesn't want to store it persistenlty in system registry).